refactor/#103 이미지 삭제를 동기방식이 아닌 이벤트를 발행해 비동기 방식으로 처리
