+++ src/main/java/com/pd/gilgeorigoreuda/common/config/S3Config.java
@@ -1,10 +1,8 @@
 package com.pd.gilgeorigoreuda.common.config;
 
-import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -21,14 +19,6 @@ public class S3Config {
 
     @Value("${cloud.aws.region.static}")
     private String region;
-
-//    @Bean
-//    public AmazonS3Client s3Client() {
-//        AWSCredentials credentials = new BasicAWSCredentials(accessKey, accessSecret);
-//        return (AmazonS3Client)AmazonS3ClientBuilder.standard()
-//                .withCredentials(new AWSStaticCredentialsProvider(credentials))
-//                .withRegion(region).build();
-//    }
 
     @Bean
     public AmazonS3 S3Client() {
