+++ src/main/java/com/pd/gilgeorigoreuda/image/domain/S3ImageDeleteEvent.java
@@ -1,0 +1,12 @@
+package com.pd.gilgeorigoreuda.image.domain;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class S3ImageDeleteEvent {
+
+    private final String imageName;
+
+}
